+++ src/components/interactive/StackedButton.jsx
@@ -4,8 +4,24 @@ import cx from 'classnames';
 import baseElement from 'util/baseElement';
 import CaretDownIcon from 'lib/icons/react/CaretDownIcon';
 
-const StackedButton = ({ Icon, Link, label, disabled, isActive, hasCaret, ...rest }) => {
+const StackedButton = ({
+  Icon,
+  Link,
+  label,
+  disabled,
+  isActive,
+  isToggled,
+  hasCaret,
+  ...rest
+}) => {
   const Element = baseElement({ href: rest.href, onClick: true, as: Link });
+
+  const ariaPressed = () => {
+    if (Element === 'button') {
+      return isToggled || isActive;
+    }
+    return undefined;
+  };
 
   return (
     <Element
@@ -18,14 +34,16 @@ const StackedButton = ({ Icon, Link, label, disabled, isActive, hasCaret, ...res
         'transition--default',
         {
           'fdsStackedButton--disabled': disabled,
-          'fdsStackedButton--active': isActive,
+          'fdsStackedButton--isActive': isActive,
+          'fdsStackedButton--isToggled': isToggled,
         }
       )}
       disabled={disabled && Element === 'button'}
+      aria-pressed={ariaPressed()}
     >
       {Icon && (
         <div className="fdsStackedButton-iconWrapper">
-          <Icon size="xs" />
+          <Icon size="s" />
           {hasCaret && <CaretDownIcon customSize={10} />}
         </div>
       )}
@@ -56,6 +74,8 @@ StackedButton.propTypes = {
   label: PropTypes.oneOfType([PropTypes.string, PropTypes.number]).isRequired,
   /** Controls the active state, which changes UI (colors) */
   isActive: PropTypes.bool,
+  /** Controls the toggled state, which changes UI (colors) */
+  isToggled: PropTypes.bool,
   /** Controls whether caret icon is displayed */
   hasCaret: PropTypes.bool,
 };